Simplify signals in the view controllers
The initial implementation of using signals relied on a confusing approach
to causing the signal to fire initially to cause data to be retrieved for the
view controllers. This has been reworked to use a more straightforward
approach using ReactiveCocoa Actions to do work and using the signal
for what its really intended, communicating that a noteworthy event has
occurred.