Convert random magic color property to Signal
The particle color was previously stored in a property and generated/stored
whenever the device orientation changed or the user navigated back to the
view. Two properties were maintained, the particle color and a flag indicating
that the scene was to be re-presented. Converting it to a signal allowed the
removal of two pieces of state; the current color and the flag.